// src/app/Model/ResumeFormModal.ts
export interface ResumeFormModal {
    FirstName: string;
    LastName: string;
    email: any;
    mobile: any;
    dob: any;
    gender: any;
    university1: any ;
    yearofpassing1: any ;
    percentage1: any ;
    university2: any ;
    yearofpassing2: any ;
    percentage2: any ;
    university3: any ;
    yearofpassing3: any ;
    percentage3: any ;
    university4: any ;
    yearofpassing4: any ;
    percentage4: any ;
    organization1: any,
      doj1: any,
      dor1: any,
    organization2: any,
      doj2: any,
      dor2: any,
      tech1: any,
      version1: any,
      lastworkedon1: any,
      rate1: any,
      tech2: any,
      version2: any,
      lastworkedon2: any,
      rate2: any,
      Objective: any,
      description: any,




}